requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
asyncpg
